eliminating compiler warnings

+++ modules/Audio/lemons_dsp/util/PitchDetector.cpp
@@ -128,17 +128,17 @@ inline float PitchDetector<SampleType>::parabolicInterpolation (int periodEstima
 	if (x0 == periodEstimate)
 	{
 		if (yinData[periodEstimate] <= yinData[x2])
-			return periodEstimate;
+            return static_cast<float> (periodEstimate);
 
-		return x2;
+        return static_cast<float> (x2);
 	}
 
 	if (x2 == periodEstimate)
 	{
 		if (yinData[periodEstimate] <= yinData[x0])
-			return periodEstimate;
+            return static_cast<float> (periodEstimate);
 
-		return x0;
+        return static_cast<float> (x0);
 	}
 
 	const auto s0 = yinData[x0];
